fix: corrigir tipo onDurationUpdate — int em vez de double

+++ frontend/lib/features/live/screening/widgets/screening_native_player_widget.dart
@@ -97,7 +97,7 @@ class _ScreeningNativePlayerWidgetState
       if (!mounted || duration == Duration.zero) return;
       ref
           .read(screeningPlayerProvider(widget.sessionId).notifier)
-          .onDurationUpdate(duration.inMilliseconds.toDouble());
+          .onDurationUpdate(duration.inMilliseconds);
     });
 
     _mkPlayer!.stream.position.listen((position) {
